new working version; more flexible
In this version, we can allow later versions of a package to satisfy
earlier versions in the sense that if we find a solution with one of the
later ones, we don't look for solutions with the earlier ones anymore.
There's no input mechanism for it yet, but we can add one. This is
useful for cases where, e.g. the first version is the "no version": if
we find some solution where the package is installed because something
else depends on it, we don't want to also consider as a solution without
the package just because the "no version" comes earlier in the ordering.
So we can say that any of normal versions dominate the "no version": we
prefer it, all other things being equal and try it first, but if we find
a solution with some version then we don't also count the "no version".