brute force: return lexigraphically optimal solutions
This fixes the tests by changing them: previously we were generating all
Pareto-optimal solutions, i.e. any solution that is not dominated by any
other solution. It turns out that's not what the algorithm I've written
computes. What it computes is the solutions that are lexicographically
optimal under some permutation of packages.