wip: this still doesn't quite work
The issue is that a later solution may already have been dominated, but
with a solution in between that wasn't already, so when we get to the
later already-dominated solution, we're looking for a new way to
dominate it, which is unnecessary and may not exist. The attempt here is
to keep the solution set in sorted order so that can't happen, but that
causes other problems because it invalidates the dominance frontier for
various levels of the call stack.