filter nodes by reachability
The concept here is that we can tell relatively easily if a node has any
possiblity of being reached by the optimal version selection algorithm
based on the following logic:

- The best version of each package is reachable
- If `v` and `w` are reachable and have a conflict then `v+1` and `w+1`
  are also reachable if they exist and belong to the same packges as `v`
  and `w` respectively

This allows us to limit the set of versions that we even consider when
searching. Of potentially even greater help, is that we can skip
unreachable versions when generating our compatibility lists, which
means that when deduplicating versions with the same compatibility
lists, we will consider two versions identical as long as their
compatibility profile is the same considering only reachable versions of
other packages. This restriction makes it much easier for two versions
to have the same compatiblity profile, allowing even more deduplication.
This is also why reachability needs to be computed before deduplicating.