add simple 2-pkg, 1-xfl example; fix bug exposed
The core solver doesn't know about requirements, just preferential order
of versions, so it will return solutions where a non-required package
is optimized by not being installed, which can force a requirement to be
unsatisfied, leading to the output of undesirable solutions, like having
no versions of any package. We address this problem by only reporting
the most satisfying solutions. If there are solutions that satisfy all
requirements, this means we won't return any that fail to do so. If no
solutions satisfy all requirements, then we'll report all the partial
solutions that satisfy as many requirements as possible.