change versions map to a lazy callback
This allows only lazily getting the versions for packages. Which is
useful when resolving within, say, a registry. We only need to decide up
front which packages are required. The versions of those packages are,
of course, looked up. Then the best version of each is considered and
its dependecies are looked up. We then go find the versions of those
packages and only consider them as we are required to do so. Doing all
of this lazily with callbacks allows us to minimize the amount of
registry parsing that needs to be done without the caller needing to
know anything about the logic of which versions are reachable when
resolving and also without the resolver needing to understand anything
about how versions, dependencies and conflicts are encoded.