sort by number of interactions
This is significantly faster than alphabetical. Oddly it's equally fast
to sort in the opposite order, i.e. from most to least interactions,
which was my initial guess for a better ordering. An interesting idea is
to use a modular permutation ordering based on the interaction graph. I
haven't thought too deeply about it, but that feels like it should work.