fix package interaction bug in find_reachable
This should have been looking at all interacting packages, not just the
ones that have a compat entry: compat entries aren't symmetrical. This
is exactly what the find_interacts helper is for, so we use that like we
already were later. To fix this I copied and modified the overall
structure of filter_redundant! which makes the code more consistent as a
pleasant side effect. In the process, I also found some places where
find_reachable was not assuming that PkgInfo has depends and compat
dicts had versions, which isn't necessarily the case. This logic was
correct in the old find_reachable, but incorrect in filter_redundant!.