simpler approach: find any valid solution, then optimize it
The idea is that the core problem specification doesn't try to include
any optimality or minimality--it just encodes what it means to be a
complete, conflict-free set of versions package requirements. To find an
optimal solutions, we find any solution and then optimize it by
improving it until it cannot be improved any more. We find the next
optimal solution by finding any solution that improves on that solution
in some aspect and then optimize that and so on.