wip: incremental optimization
This doesn't quite work for a subtle reason: when you force a better
version of some package it might still get deleted during the
minimization step which doesn't use the SAT solver--it just deletes any
packages that are not reachable. This is probably not a new problem!
Previous approaches used the SAT solver to do minimization, which means
that they would not fully minimize their solutions in such cases because
minimization would also obey the added SAT clauses. The solution is to
only force better versions of the reqs at the outermost level and do the
same thing recursively inside. That's a significant refactoring.