wip: find a solution for each optimal choice of reqs
This this doesn't find every optimal solution, only an optimal solution
for each possible choice of top-level requirement versions. If we want
to find all the pareto-optimal solutions, we need to do this
recursively; even then, we won't find all possible optimal solutions
since it can never find a solution that sacrifices a top-level package
for the sake of an indirect dependency. But that's probably a good thing
anyway since those solutions would be a bit weird for a user to choose.
In a pinch, they can be forced by putting compat restrictions on an
indirect dependency.