load_pkg_info: avoid O(n^2) population of conflict bits
The old code had to iterate through every pair of versions of every pair
of interacting packages and check for compatibility. That was slow. This
version lets the structure of the compat data drive instead and so only
looks at combinations that might actually have a conflict. This takes
time to populate the conflict bits from 6s to 0.2s, which is a huge
reduction. Total time for loading all package info goes from 8.1s down
to 2.3s. Moreover, after this that's no longer the bottleneck. Now it's the filtering step, which I can tackle optimization of next.