Fix subtle solution ordering bug: tweak order with unsat requirements
We have a subtle (but fixable) problem here: we include solutions that
don't even satisfy all requirements in our potential solutions, which
kind of works just by considering the non-version as the worst version
for necessary packages. However, this means that an incomplete but valid
solution may not be dominated by any of the complete optimal solutions.
It doesn't dominate them either since they are better in the dimension
of having actual versions for all requirements. But it can dominate them
by choosing a better version of some required package in isolation, at
the cost of not being able to satisfy all the requirements. However,
this is not a solution we'd ever want to choose unless we can't fully
satisfy all our requirements.

Fix: compare satisfaction sets before even considering how good the
versions chosen are. Any solution that satisfies a strict superset of
necessary packages than another solution is strictly better. We only
look at the actual versions chosen if the satisfied set is the same. If
the satisfied sets are different but neither contains the other, then
the solutions are incomparable.