return maximal incomplete solutions when unsatisfiable
This works for smaller solutions but for "resolving the world" we're
going to need a different approach: this produces too many partial
solutions--it seems like an exponential number. It would be appealing to
find partial solutions that maximize the number of satisfied
requirements, but it's a little unclear how to do that with a plain SAT
solver, as opposed to a ILP optimizer.