slices: use initial slices to compute true compat graph then redo slices
The comment explains this in more detail, but computing the true
pairwise compatibility graph between the best versions of packages is
not naively feasible O(N^2), but we can use the explicit conflict graph
to find an initial coloring, expand that into full slices, and then use
that to essentially construct a block matrix of which packages don't
ever appear in a slice together. We can efficiently check satisfiability
of the block-diagonal pairs and compute a true pairwise compat graph.
From there, we can use that graph to compute better colors and slices.