slices: implement DSatur graph coloring algorithm
Not only is this much faster, but it turns out to do a strictly better
job than RLF does for this problem. With the explicit conflict graph it
produces one fewer colors/slices (6 vs 7) and with the implicit conflict
graph it also produces one fewer cloros/slices (5 vs 6). I was
considering calling both algorithms for the final step and using results
of the one that produced fewer colors, but it seems like just using
DSatur may be the way to go.