slices: Dsatur: use new variable per color instead of whole SAT instance
When you add a new color, add a variable for that color and when new
vertices are added to the color, add a clause that the color implies the
vertex. To check if a new vertex can be added to a color, assume the
color and the vertex and check satisfiability.