slices: use ordering to pick specific solutions
The concept was to pick consistent lexicographically-first solutions by
package popularity to try to ensure that our final slices are included
in the package/version vertices that we already have, but that's still
not guaranteed by this, so more work is needed. Looks like we'll need to
iteratively expand the vertex set until it reaches a fixed point.