Handle weakdeps while loading registries
Amazingly, this required only one very tiny tweak to the Resolver logic.
Previously we assumed that any package with a key in a compat dictionary
must be reachable and thus in data, but that's not the case with weak
deps. The tweak is simple: if a package appears in some compat dict but
isn't in data, then we can just ignore it.