Julia needs to compat-lock stdlibs; prefer stdlibs over registered
It's not sufficient for a stdlib version to compat-lock with its Julia
version because then you can resolve a different stdlib version for that
Julia version, which isn't actually feasible.

When generating a manifest if a package is a stdlib for the chosen Julia
version, then we need to prefer the stdlib version even if that same
version appears in a registry.