After resolving, re-find the right HistoricalStdlibVersions entry
This shouldn't be necessary, but HistoricalStdlibVersions assigns the
same version number to stdlib versions with different deps. Yes, this
is totally wrong and should not happen. And yet it does. So we resolve
taking those version numbers at face value and taking the union of the
deps, but then when generating the manifest, we look up the variant of
that stdlib version that is specific to the Julia version we resolved.

This eliminates all but one tiny difference when resolving a project
with DifferentialEquations as its sole dependency and allows the project
to be loaded. The one difference is that we resolve OpenLibm_jll 0.8.1+2
instead of 0.8.5+0. I suspect the difference is due to extra deps being
imagined for that stdlib JLL (JLLWrappers).